Intial Commit (take 2)
This is the inital commit to the Discord display branch. I haven't tested this yet with all of the newer changes/structure that have changed but it should be fine. Once I have time, I will do testing and commit any additional changes but I wanted to get something pushed sooner rather than later.

Removed previous branch/commit since I forgot that I hadn't setup my replacement script for config.php in github desktop. Easiest way of dealing with it was to remove the branch and recreate it.